% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_plots.R
\name{all_possible_plots_df}
\alias{all_possible_plots_df}
\title{Produce a table of all possible plots
This table will have columns:}
\usage{
all_possible_plots_df(loc)
}
\arguments{
\item{loc}{A two-letter code for the location}
}
\description{
- category
    - shortname 
    - name
    - plot
}
\details{
This function does not produce the plots, but simply leaves an empty list()
item for each the place of each plot. Plots are produced by the calib_plots
function.
}
