% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scripts.R
\name{calib_plt_pct_cases_nusb}
\alias{calib_plt_pct_cases_nusb}
\title{Plot Percent of TB Cases in Non-US-Born Individuals}
\usage{
calib_plt_pct_cases_nusb(loc)
}
\description{
Plot Percent of TB Cases in Non-US-Born Individuals
}
