% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scripts.R
\name{calib_plt_pct_cases_nusb_recent}
\alias{calib_plt_pct_cases_nusb_recent}
\title{Plot Percent of non-US born TB cases from recent immigrants (<2 yrs)}
\usage{
calib_plt_pct_cases_nusb_recent(loc)
}
\description{
Plot Percent of non-US born TB cases from recent immigrants (<2 yrs)
}
