% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scripts.R
\name{calib_plt_pop_by_age_nat}
\alias{calib_plt_pop_by_age_nat}
\title{Plot Total Population by Age and Nativity}
\usage{
calib_plt_pop_by_age_nat(loc)
}
\description{
Plot Total Population by Age and Nativity
}
