% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scripts.R
\name{calib_plt_pop_by_nat_over_time}
\alias{calib_plt_pop_by_nat_over_time}
\title{Plot Total Population by Nativity over Time}
\usage{
calib_plt_pop_by_nat_over_time(loc)
}
\description{
Plot Total Population by Nativity over Time
}
