% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scripts.R
\name{calib_plt_us_ltbi_by_age}
\alias{calib_plt_us_ltbi_by_age}
\title{Plot LTBI Prevalence in Non-US-Born Individuals by Age}
\usage{
calib_plt_us_ltbi_by_age(loc)
}
\description{
Plot LTBI Prevalence in Non-US-Born Individuals by Age
}
