% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Check that Data Exist in MITUS for Each Possible Plot}
\usage{
data_check(loc, possible_plots)
}
\arguments{
\item{loc}{A two-letter code for the location}

\item{possible_plots}{A vector of possible plot shortnames}
}
\value{
A logical vector indicating if data exist for each of possible_plots
}
\description{
Check that Data Exist in MITUS for Each Possible Plot
}
