% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MITUSCalibPlots-package}
\alias{MITUSCalibPlots}
\alias{MITUSCalibPlots-package}
\title{Generate Calibration Plots from MITUS for Tabby2}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
